Add files via upload
Dataset= csv file, the .jl file generates this dataset. Other datasets can be made changing the seed.
The distribution chosen based on the KS-test is saved in the best_ks_distribution.txt file, and the .jl code's output is in the ostat_output_log.txt file, the rest are plot images.